refactor(survey): decompose the survey manager god-file by role (WS-D)
Split internal/wifi/survey/survey.go (893 LOC) into four role-based
files in the same package — pure move, no behavior change:

- survey.go    domain model: Type/Status/band consts, FloorPlan, Floor,
               PassiveSample(+CalculateAggregations), ActiveSample,
               ThroughputSample, SamplePoint, Survey(+helpers),
               APLocation, ClientLocation, PassFailCriterion
- manager.go   Manager construction + survey CRUD/lifecycle + settings,
               imported-data, and sample addition
- floors.go    floor and floor-plan operations
- store.go     on-disk persistence (save/load/migrate)

All four are now under the 600-line size-gate threshold. Declaration
sets and code-line multisets verified identical to the original. Two
malformed dangling doc-comment lines on the persistence funcs were
tidied while moving (comment-only, no code change).

Related to #187
