+++ internal/wifi/survey/survey.go
@@ -1,18 +1,8 @@
 package survey
 
 import (
-	"encoding/json"
-	"errors"
-	"fmt"
-	"os"
-	"path/filepath"
-	"sync"
 	"time"
 
-	"github.com/google/uuid"
-
-	"github.com/MustardSeedNetworks/seed/internal/diagnostics/iperf"
-	"github.com/MustardSeedNetworks/seed/internal/logging"
 	"github.com/MustardSeedNetworks/seed/internal/wifi"
 )
 
@@ -296,598 +286,4 @@ func (s *Survey) GetAllSamples() []*SamplePoint {
 	// Include legacy samples if present
 	samples = append(samples, s.Samples...)
 	return samples
-}
-
-// Manager manages WiFi site surveys.
-type Manager struct {
-	mu           sync.RWMutex
-	surveys      map[string]*Survey // key is survey ID
-	storagePath  string
-	wifiScanner  *wifi.Scanner
-	wifiManager  *wifi.Manager
-	iperfManager *iperf.Manager
-}
-
-// NewManager creates a new survey manager.
-func NewManager(
-	storagePath string,
-	wifiScanner *wifi.Scanner,
-	wifiManager *wifi.Manager,
-	iperfManager *iperf.Manager,
-) *Manager {
-	return &Manager{
-		surveys:      make(map[string]*Survey),
-		storagePath:  storagePath,
-		wifiScanner:  wifiScanner,
-		wifiManager:  wifiManager,
-		iperfManager: iperfManager,
-	}
-}
-
-// CreateSurvey creates a new survey with a default floor.
-func (m *Manager) CreateSurvey(name, description, iface string, surveyType Type) (*Survey, error) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	now := time.Now()
-
-	// Create default first floor
-	defaultFloor := &Floor{
-		ID:        uuid.New().String(),
-		Name:      "Floor 1",
-		Level:     1,
-		Samples:   make([]*SamplePoint, 0),
-		CreatedAt: now,
-		UpdatedAt: now,
-	}
-
-	survey := &Survey{
-		ID:            uuid.New().String(),
-		Name:          name,
-		Description:   description,
-		SurveyType:    surveyType,
-		Status:        StatusCreated,
-		CreatedAt:     now,
-		UpdatedAt:     now,
-		Interface:     iface,
-		Floors:        []*Floor{defaultFloor},
-		ActiveFloorID: defaultFloor.ID,
-	}
-
-	m.surveys[survey.ID] = survey
-
-	// Persist to disk
-	if err := m.saveSurvey(survey); err != nil {
-		return nil, fmt.Errorf("failed to save survey: %w", err)
-	}
-
-	return survey, nil
-}
-
-// GetSurvey retrieves a survey by ID.
-func (m *Manager) GetSurvey(id string) (*Survey, error) {
-	m.mu.RLock()
-	defer m.mu.RUnlock()
-
-	survey, exists := m.surveys[id]
-	if !exists {
-		return nil, fmt.Errorf("survey not found: %s", id)
-	}
-
-	return survey, nil
-}
-
-// ListSurveys returns all surveys.
-func (m *Manager) ListSurveys() []*Survey {
-	m.mu.RLock()
-	defer m.mu.RUnlock()
-
-	surveys := make([]*Survey, 0, len(m.surveys))
-	for _, survey := range m.surveys {
-		surveys = append(surveys, survey)
-	}
-
-	return surveys
-}
-
-// DeleteSurvey deletes a survey.
-func (m *Manager) DeleteSurvey(id string) error {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	if _, exists := m.surveys[id]; !exists {
-		return fmt.Errorf("survey not found: %s", id)
-	}
-
-	// Delete from disk
-	filename := filepath.Join(m.storagePath, fmt.Sprintf("%s.json", id))
-	if err := os.Remove(filename); err != nil && !os.IsNotExist(err) {
-		return fmt.Errorf("failed to delete survey file: %w", err)
-	}
-
-	delete(m.surveys, id)
-
-	return nil
-}
-
-// UpdateFloorPlan updates the floor plan for the active floor (or specified floor).
-func (m *Manager) UpdateFloorPlan(id string, floorPlan *FloorPlan) error {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	survey, exists := m.surveys[id]
-	if !exists {
-		return fmt.Errorf("survey not found: %s", id)
-	}
-
-	// Find the active floor
-	floor := survey.GetActiveFloor()
-	if floor == nil {
-		return fmt.Errorf("no active floor set for survey: %s", id)
-	}
-
-	floor.FloorPlan = floorPlan
-	floor.UpdatedAt = time.Now()
-	survey.UpdatedAt = time.Now()
-
-	return m.saveSurvey(survey)
-}
-
-// UpdateFloorPlanByFloorID updates the floor plan for a specific floor.
-func (m *Manager) UpdateFloorPlanByFloorID(surveyID, floorID string, floorPlan *FloorPlan) error {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	survey, exists := m.surveys[surveyID]
-	if !exists {
-		return fmt.Errorf("survey not found: %s", surveyID)
-	}
-
-	floor := survey.GetFloorByID(floorID)
-	if floor == nil {
-		return fmt.Errorf("floor not found: %s", floorID)
-	}
-
-	floor.FloorPlan = floorPlan
-	floor.UpdatedAt = time.Now()
-	survey.UpdatedAt = time.Now()
-
-	return m.saveSurvey(survey)
-}
-
-// StartSurvey starts a survey.
-func (m *Manager) StartSurvey(id string) error {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	survey, exists := m.surveys[id]
-	if !exists {
-		return fmt.Errorf("survey not found: %s", id)
-	}
-
-	if survey.Status == StatusInProgress {
-		return errors.New("survey already in progress")
-	}
-
-	survey.Status = StatusInProgress
-	survey.UpdatedAt = time.Now()
-
-	return m.saveSurvey(survey)
-}
-
-// PauseSurvey pauses a survey.
-func (m *Manager) PauseSurvey(id string) error {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	survey, exists := m.surveys[id]
-	if !exists {
-		return fmt.Errorf("survey not found: %s", id)
-	}
-
-	survey.Status = StatusPaused
-	survey.UpdatedAt = time.Now()
-
-	return m.saveSurvey(survey)
-}
-
-// CompleteSurvey completes a survey.
-func (m *Manager) CompleteSurvey(id string) error {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	survey, exists := m.surveys[id]
-	if !exists {
-		return fmt.Errorf("survey not found: %s", id)
-	}
-
-	survey.Status = StatusCompleted
-	survey.UpdatedAt = time.Now()
-
-	return m.saveSurvey(survey)
-}
-
-// UpdateSurveySettings updates survey settings (only when survey is in created state).
-func (m *Manager) UpdateSurveySettings(
-	id string,
-	surveyType Type,
-	iperfServer string,
-	testDuration int,
-) error {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	survey, exists := m.surveys[id]
-	if !exists {
-		return fmt.Errorf("survey not found: %s", id)
-	}
-
-	if survey.Status != StatusCreated {
-		return errors.New("cannot update settings after survey has started")
-	}
-
-	// Validate survey type
-	switch surveyType {
-	case TypePassive, TypeActive, TypeThroughput:
-		survey.SurveyType = surveyType
-	default:
-		return fmt.Errorf("invalid survey type: %s", surveyType)
-	}
-
-	// Validate test duration
-	if testDuration < 1 {
-		testDuration = defaultTestDurationSec
-	}
-	if testDuration > maxTestDurationSec {
-		testDuration = maxTestDurationSec
-	}
-	survey.TestDuration = testDuration
-	survey.IperfServer = iperfServer
-	survey.UpdatedAt = time.Now()
-
-	return m.saveSurvey(survey)
-}
-
-// ImportedDataUpdate carries the slices a caller wants to replace on a survey.
-// Each field is optional; nil means "leave as-is". An empty (non-nil) slice
-// clears the corresponding survey field, which is the right behaviour when an
-// importer wants to fully replace the existing list.
-type ImportedDataUpdate struct {
-	APLocations      []APLocation
-	ClientLocations  []ClientLocation
-	PassFailCriteria []PassFailCriterion
-}
-
-// UpdateImportedData replaces the survey's imported-data fields (#727).
-// Callers (most notably the AirMapper import flow) use this to persist the
-// AP/client placements and pass-fail criteria that come back from the parser
-// so the data survives a page reload.
-func (m *Manager) UpdateImportedData(id string, update ImportedDataUpdate) error {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	survey, exists := m.surveys[id]
-	if !exists {
-		return fmt.Errorf("survey not found: %s", id)
-	}
-
-	if update.APLocations != nil {
-		survey.APLocations = update.APLocations
-	}
-	if update.ClientLocations != nil {
-		survey.ClientLocations = update.ClientLocations
-	}
-	if update.PassFailCriteria != nil {
-		survey.PassFailCriteria = update.PassFailCriteria
-	}
-	survey.UpdatedAt = time.Now()
-
-	return m.saveSurvey(survey)
-}
-
-// AddSample adds a measurement sample to the active floor of a survey.
-func (m *Manager) AddSample(id string, x, y int, sampleData any) error {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	survey, exists := m.surveys[id]
-	if !exists {
-		return fmt.Errorf("survey not found: %s", id)
-	}
-
-	if survey.Status != StatusInProgress {
-		return errors.New("survey not in progress")
-	}
-
-	floor := survey.GetActiveFloor()
-	if floor == nil {
-		return fmt.Errorf("no active floor set for survey: %s", id)
-	}
-
-	sample := &SamplePoint{
-		X:          x,
-		Y:          y,
-		Timestamp:  time.Now(),
-		SampleData: sampleData,
-	}
-
-	floor.Samples = append(floor.Samples, sample)
-	floor.UpdatedAt = time.Now()
-	survey.UpdatedAt = time.Now()
-
-	return m.saveSurvey(survey)
-}
-
-// AddSampleToFloor adds a measurement sample to a specific floor.
-func (m *Manager) AddSampleToFloor(surveyID, floorID string, x, y int, sampleData any) error {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	survey, exists := m.surveys[surveyID]
-	if !exists {
-		return fmt.Errorf("survey not found: %s", surveyID)
-	}
-
-	if survey.Status != StatusInProgress {
-		return errors.New("survey not in progress")
-	}
-
-	floor := survey.GetFloorByID(floorID)
-	if floor == nil {
-		return fmt.Errorf("floor not found: %s", floorID)
-	}
-
-	sample := &SamplePoint{
-		X:          x,
-		Y:          y,
-		Timestamp:  time.Now(),
-		SampleData: sampleData,
-	}
-
-	floor.Samples = append(floor.Samples, sample)
-	floor.UpdatedAt = time.Now()
-	survey.UpdatedAt = time.Now()
-
-	return m.saveSurvey(survey)
-}
-
-// AddFloor adds a new floor to a survey.
-func (m *Manager) AddFloor(surveyID, name string, level int) (*Floor, error) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	survey, exists := m.surveys[surveyID]
-	if !exists {
-		return nil, fmt.Errorf("survey not found: %s", surveyID)
-	}
-
-	now := time.Now()
-	floor := &Floor{
-		ID:        uuid.New().String(),
-		Name:      name,
-		Level:     level,
-		Samples:   make([]*SamplePoint, 0),
-		CreatedAt: now,
-		UpdatedAt: now,
-	}
-
-	survey.Floors = append(survey.Floors, floor)
-	survey.UpdatedAt = now
-
-	if err := m.saveSurvey(survey); err != nil {
-		return nil, err
-	}
-
-	return floor, nil
-}
-
-// UpdateFloor updates floor metadata (name, level).
-func (m *Manager) UpdateFloor(surveyID, floorID, name string, level int) error {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	survey, exists := m.surveys[surveyID]
-	if !exists {
-		return fmt.Errorf("survey not found: %s", surveyID)
-	}
-
-	floor := survey.GetFloorByID(floorID)
-	if floor == nil {
-		return fmt.Errorf("floor not found: %s", floorID)
-	}
-
-	floor.Name = name
-	floor.Level = level
-	floor.UpdatedAt = time.Now()
-	survey.UpdatedAt = time.Now()
-
-	return m.saveSurvey(survey)
-}
-
-// DeleteFloor removes a floor from a survey.
-func (m *Manager) DeleteFloor(surveyID, floorID string) error {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	survey, exists := m.surveys[surveyID]
-	if !exists {
-		return fmt.Errorf("survey not found: %s", surveyID)
-	}
-
-	// Don't allow deletion of the last floor
-	if len(survey.Floors) <= 1 {
-		return errors.New("cannot delete the last floor")
-	}
-
-	// Find and remove the floor
-	for i, floor := range survey.Floors {
-		if floor.ID == floorID {
-			survey.Floors = append(survey.Floors[:i], survey.Floors[i+1:]...)
-
-			// If we deleted the active floor, switch to the first remaining floor
-			if survey.ActiveFloorID == floorID {
-				survey.ActiveFloorID = survey.Floors[0].ID
-			}
-
-			survey.UpdatedAt = time.Now()
-			return m.saveSurvey(survey)
-		}
-	}
-
-	return fmt.Errorf("floor not found: %s", floorID)
-}
-
-// SetActiveFloor sets the active floor for data collection.
-func (m *Manager) SetActiveFloor(surveyID, floorID string) error {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	survey, exists := m.surveys[surveyID]
-	if !exists {
-		return fmt.Errorf("survey not found: %s", surveyID)
-	}
-
-	// Verify floor exists
-	floor := survey.GetFloorByID(floorID)
-	if floor == nil {
-		return fmt.Errorf("floor not found: %s", floorID)
-	}
-
-	survey.ActiveFloorID = floorID
-	survey.UpdatedAt = time.Now()
-
-	return m.saveSurvey(survey)
-}
-
-// GetFloors returns all floors for a survey.
-func (m *Manager) GetFloors(surveyID string) ([]*Floor, error) {
-	m.mu.RLock()
-	defer m.mu.RUnlock()
-
-	survey, exists := m.surveys[surveyID]
-	if !exists {
-		return nil, fmt.Errorf("survey not found: %s", surveyID)
-	}
-
-	return survey.Floors, nil
-}
-
-// GetFloor returns a specific floor.
-func (m *Manager) GetFloor(surveyID, floorID string) (*Floor, error) {
-	m.mu.RLock()
-	defer m.mu.RUnlock()
-
-	survey, exists := m.surveys[surveyID]
-	if !exists {
-		return nil, fmt.Errorf("survey not found: %s", surveyID)
-	}
-
-	floor := survey.GetFloorByID(floorID)
-	if floor == nil {
-		return nil, fmt.Errorf("floor not found: %s", floorID)
-	}
-
-	return floor, nil
-}
-
-// saveSurvey persists a survey to disk.
-//
-
-func (m *Manager) saveSurvey(survey *Survey) error {
-	// Ensure storage directory exists with restrictive permissions
-	if err := os.MkdirAll(m.storagePath, 0o750); err != nil {
-		return fmt.Errorf("failed to create storage directory: %w", err)
-	}
-
-	filename := filepath.Join(m.storagePath, fmt.Sprintf("%s.json", survey.ID))
-	data, err := json.MarshalIndent(survey, "", "  ")
-	if err != nil {
-		return fmt.Errorf("failed to marshal survey: %w", err)
-	}
-
-	if writeErr := os.WriteFile(filename, data, 0o600); writeErr != nil {
-		return fmt.Errorf("failed to write survey file: %w", writeErr)
-	}
-
-	return nil
-}
-
-// LoadSurveys loads all surveys from disk and auto-migrates legacy single-floor surveys.
-// Fixes #872: Added timeout protection for file operations (30 second limit).
-func (m *Manager) LoadSurveys() error {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	// Create storage directory if it doesn't exist with restrictive permissions
-	if err := os.MkdirAll(m.storagePath, 0o750); err != nil {
-		return fmt.Errorf("failed to create storage directory: %w", err)
-	}
-
-	// Read all JSON files
-	files, err := filepath.Glob(filepath.Join(m.storagePath, "*.json"))
-	if err != nil {
-		return fmt.Errorf("failed to list survey files: %w", err)
-	}
-
-	// Fixes #872: Track start time to prevent hanging on slow filesystem
-	startTime := time.Now()
-	const loadTimeout = 30 * time.Second
-
-	for _, file := range files {
-		// Check timeout to prevent hanging on slow/unresponsive filesystem (fixes #872)
-		if time.Since(startTime) > loadTimeout {
-			logging.GetLogger().
-				Warn("Survey loading timeout reached, some surveys may not be loaded",
-					"loaded_count", len(m.surveys),
-					"elapsed", time.Since(startTime))
-			break
-		}
-
-		// file comes from filepath.Glob with our controlled pattern, not user input
-		data, readErr := os.ReadFile(file)
-		if readErr != nil {
-			continue // Skip files that can't be read
-		}
-
-		var survey Survey
-		if unmarshalErr := json.Unmarshal(data, &survey); unmarshalErr != nil {
-			continue // Skip invalid JSON
-		}
-
-		// Auto-migrate legacy single-floor surveys to multi-floor format
-		if MigrateToMultiFloor(&survey) {
-			// Save the migrated survey back to disk
-			if saveErr := m.saveSurveyUnlocked(&survey); saveErr != nil {
-				logging.GetLogger().
-					Warn("Failed to save migrated survey", "survey_id", survey.ID, "error", saveErr)
-			}
-		}
-
-		m.surveys[survey.ID] = &survey
-	}
-
-	return nil
-}
-
-// saveSurveyUnlocked persists a survey to disk without acquiring a lock.
-// This is used internally when the lock is already held.
-//
-
-func (m *Manager) saveSurveyUnlocked(survey *Survey) error {
-	// Ensure storage directory exists with restrictive permissions
-	if err := os.MkdirAll(m.storagePath, 0o750); err != nil {
-		return fmt.Errorf("failed to create storage directory: %w", err)
-	}
-
-	filename := filepath.Join(m.storagePath, fmt.Sprintf("%s.json", survey.ID))
-	data, err := json.MarshalIndent(survey, "", "  ")
-	if err != nil {
-		return fmt.Errorf("failed to marshal survey: %w", err)
-	}
-
-	if writeErr := os.WriteFile(filename, data, 0o600); writeErr != nil {
-		return fmt.Errorf("failed to write survey file: %w", writeErr)
-	}
-
-	return nil
 }
